Add files via upload
This project is based on Mammographic Mass Data, which is available from the cloud-native data catalogue called data.world website. 
Link for the dataset: 
https://data.world/julio/mammographic-masses 
The given dataset was donated to the data.world by Matthias Elter from the Fraunhofer Institute for Integrated Circuits (IIS) in 2007, and it is available for download in the CSV format.